fix some formatting and consistency issues, implement the first part of the driver.

+++ src/Bank.java
@@ -5,27 +5,27 @@
  * A file that does stuff.
  */
 public class Bank {
-    public long publicKey = 1;
-    private long privateKey = 1;
-    public long modulus = 1;
-    
-    private Bank() {
-	
-    }
-    
-    public static void seedPrng(long seed) {
-	
-    }
-    
-    public static long selectUnopenedMoneyOrder(long maximum) {
-	return 1;
-    }
-    
-    public static boolean verifyAndSignMoneyOrders(long unopenedOrder) {
-	return true;
-    }
-    
-    public static boolean verifyMerchantMoneyOrder() {
-	return true;
-    }
+	public long publicKey = 1;
+	private long privateKey = 1;
+	public long modulus = 1;
+
+	private Bank() {
+
+	}
+
+	public static void seedPrng(long seed) {
+
+	}
+
+	public static long selectUnopenedMoneyOrder(int maximum) {
+		return 1;
+	}
+
+	public static boolean verifyAndSignMoneyOrders(int unopenedOrder) {
+		return true;
+	}
+
+	public static boolean verifyMerchantMoneyOrder() {
+		return true;
+	}
 }
